Fix unit test

+++ x/evm/module_test.go
@@ -48,7 +48,7 @@ func TestModuleExportGenesis(t *testing.T) {
 func TestConsensusVersion(t *testing.T) {
 	k, _ := testkeeper.MockEVMKeeper()
 	module := evm.NewAppModule(nil, k)
-	assert.Equal(t, 5, module.ConsensusVersion())
+	assert.Equal(t, uint64(5), module.ConsensusVersion())
 }
 
 func TestABCI(t *testing.T) {
